Refactor StatisticalDistrictListComponent to fix errors
