+++ src/views/AreaView/components/StatisticalDistrictList/index.js
@@ -1,5 +1,5 @@
 import { withStyles } from '@mui/styles';
-import StatisticalDistrictList from './StatisticalDistrictList';
+import StatisticalDistrictList from './StatisticalDistrictListComponent';
 import styles from '../../styles';
 
 export default withStyles(styles)(StatisticalDistrictList);
